Coherence analyses

ContinuousWavelet.coherence — Method

coherence(wt::ContinuousWaveletTransform,
          X::AbstractArray{Float64, 2};
          nsurrogate=0, α=0.05)

Implements the self-coherence of a set of time-series. The self-coherence is defined as

\[ C(X)(b,a) = \frac{\left\vert E\left[\mathcal{W}[X](b,a)\right]\right\vert^2}{E\left[\left\vert\mathcal{W}[X](b,a)\right\vert^2\right]} = \frac{\left\vert\mathcal{W}[E\left[X\right]](b,a)\right\vert^2}{E\left[\left\vert\mathcal{W}[X](b,a)\right\vert^2\right]}\]

This measure can be interpreted as how coherent a process is from realization to realization. Typical applications are the analyses of EEG event-related potentials and similar time-series of repeated experiments.

Arguments

wt: Specifies the wavelet transform to use.

X: Specifies the time-series to analyze. Each single time-series (trial) is stored as a column in X.

'nsurrogate': A keyword argument specifying the number of surrogate samples to generate from the given time-series for the estimation of the point-wise significant test. By default it is 0 and no surrogate samples are generated.

α: A keyword arguement specifying the significance level for the point-wise significance test. By default set to α=0.05.

ContinuousWavelet.coherence — Method

coherence(wt::ContinuousWaveletTransform,
          X::AbstractArray{Float64, 2}, Y::AbstractArray{Float64, 2};
          nsurrogate=0, α=0.05)

Canonical wavelet coherence as defined by

\[ C(X)(b,a) = \frac{E\left[\mathcal{W}[X](b,a)\,\mathcal{W}[Y]^{*}(b,a)\right]}{ E\left[\left\vert\mathcal{W}[X](b,a)\right\vert\,\left\vert\mathcal{W}[Y](b,a)\right\vert\right]}\]

That is the normalized expectation value of the product of the two wavelet transformed, where z^* is the complex conjugate.

In contrast to the self-coherence, this analysis does estimate the coherence of a process to itself but the coherence between two observed processes. It is therefore necessary that the two sets of time-series X and Y are of the same shape. That is, each trial of X and Y must consists of the same number of samples (rows) and both sets must contain the same number of trials (columns).

Arguments

wt: Specifies the wavelet transform to use.

X: Specifies one time-series to analyze. Each single time-series (trial) is stored as a column in X. Must be of same shape as Y.

Y: Specifies one time-series to analyze. Each single time-series (trial) is stored as a column in Y. Must be of same shape as X.

'nsurrogate': A keyword argument specifying the number of surrogate samples to generate from the given time-series for the estimation of the point-wise significant test. By default it is 0 and no surrogate samples are generated.

α: A keyword arguement specifying the significance level for the point-wise significance test. By default set to α=0.05.

Coherence analysis result

ContinuousWavelet.ContinuousWaveletCoherence — Type

Represents a continuous wavelet coherence analysis.

Fields:

coh: The actual coherence as a complex value. The layout of this array follows the layout of ContinuousWaveletTransformed. That is, each voice is stored in a column, starting with the smallest scale.

scales: The vector of scales at which the wavelet transform is performed.

wavelet: The wavelet being used.

α: The point-wise empirical signigicance level of the coherence analysis obtained by means of surrogate data. Holds NaN if no skipped.

Missing docstring.

Missing docstring for Plots.contourf(A::ContinuousWavelet.ContinuousWaveletCoherence; kw...). Check Documenter's build log for details.