Continuous Wavelets

The ContinuousWavelet library implements (for now) only two mother wavelets.

ContinuousWavelet.CauchyWavelet — Type

The Cauchy or Paul wavelet.

In contrast to the MorletWavelet, the Cauchy or Paul wavelet is a proper wavelet with a similar good localization in time and scale.

There are different definitions of the Cauchy wavelet around. Here one is implemented where the center frequency is always 1 irrespective of the value of α:

\[ g(t) = h(t) = \left(1-i\,\frac{2\pi\,t}{\alpha}\right)^{-(1+\alpha)}\,,\]

and its reproducing kernel

\[ P_{g,h}(b, a) = \Gamma(2\alpha+1)\,a^{\alpha+1}\,\left(1+a-\frac{i\,b}{a}\right)^{-(2\alpha+1)}\,.\]

ContinuousWavelet.MorletWavelet — Type

The Morlet wavelet.

\[ g(t) = \sqrt{\frac{\delta}{2\pi}}\exp(2\pi\,i\,t-t^2\,\delta)\,,\]

where δ specifies the time-frequency resolution of the wavelet.

All mother wavelets are derived from the abstract type

ContinuousWavelet.GenericContinuousWavelet — Type

Represents the base-class of all continuous wavelets.

Methods

ContinuousWavelet.eval_analysis — Function

eval_analysis(wav::ContinuousWavelet.GenericContinuousWavelet, t::Float64)

Evaluates the mother wavelet at timepoint t.

ContinuousWavelet.eval_synthesis — Function

eval_synthesis(wav::ContinuousWavelet.GenericContinuousWavelet, t::Float64)

Evaluates the mother synthesis wavelet at timepoint t.

ContinuousWavelet.eval_repkern — Function

eval_repkern(wav::ContinuousWavelet.GenericContinuousWavelet, a::Float64, b::Float64)

Evaluates the reproducing kernel at scale a and time-point b.

ContinuousWavelet.cutoff_time — Function

cutoff_time(wav::ContinuousWavelet.GenericContinuousWavelet)

Returns the cut-off time-point for the given wavelet at scale 1.

ContinuousWavelet.cutoff_freq — Function

cutoff_freq(wav::ContinuousWavelet.GenericContinuousWavelet)

Returns the cut-off frequency for the given wavelet at scale 1.